Launcher: Fix for a null pointer exception while renaming a folder
In a particular scenario, when user tries to rename a folder, and before
the user selects either OK or CANCEL button on the rename dialog, the
Launcher process gets recreated and is in a pending state of renaming
the folder which is made note of by setting a flag.  In onCreate()
function that gets called, the restoreState() call retreives the
folder by querying the launcher model. It is the startLoaders()
call [in onCreate() of Launcher.java] that actually prepares the
Launcher model, but since this is called after the restoreState(), the
mFolders object of LauncherModel still remain empty resulting in null
pointer exception.

CRs-Fixed:188784