First try to create a fallback for the DragLayer memory issues
When users longpress anything, the DragLayer tries to create the drawing cache of that View,
and create a BitMap copy of the drawing cache to move it around the screen.
Under low memory conditions, the drawing cache cannot be generated by android, so it throws nasty NPEs
while trying to access that drawing cache.

I've created a fallback to draw just roundRects when the cache cannot be generated.
TODO: If the cache can be generated, still could throw an OOM Error whith the createBitmap method....
I wish i could catch it to make the fallback work on that case too....